requirements.txt
google-adk==0.1.0
supabase>=2.13.0
litellm
openai
duckduckgo-search
google-generativeai 